refactor: 컴포넌트 위치 변경 및 구조변경
